Weather App
Our app is going to make an HTTP call to a weather API to retrieve data whenever a user enters a city into a form. We are going to use conditional logic, styling, spinners, and error handling to improve the user experience as we wait for that HTTP call to return our data. Before we can get started, let’s go over a few prerequisites.